rise_case/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
joblib==1.3.2
langchain==0.0.279
langsmith==0.0.33
loguru==0.7.0
marshmallow==3.20.1
mmh3==3.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pinecone-client==2.2.2
pinecone-text==0.5.3
pydantic==2.3.0
pydantic-core==2.6.3
pypdf==3.15.4
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==2.0.4
uvicorn==0.23.2
wget==3.2
yarl==1.9.2